requirements.txt
opencv-contrib-python>=4.10.0
numpy>=1.26.0
pillow>=10.0.0
